cleaned up

+++ incs/libasm.h
@@ -6,7 +6,7 @@
 /*   By: jabenjam <jabenjam@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/19 17:12:13 by jabenjam          #+#    #+#             */
-/*   Updated: 2020/09/19 17:13:23 by jabenjam         ###   ########.fr       */
+/*   Updated: 2020/09/21 19:53:04 by jabenjam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@
 size_t		ft_strlen(const char *str);
 char		*ft_strcpy(char *dst, char *src);
 int			ft_strcmp(const char *s1, const char *s2);
-int			ft_write(int fildes, const void *buff, size_t nbyte);
+ssize_t		ft_write(int fildes, const void *buff, size_t nbyte);
 ssize_t		ft_read(int fildes, void *buf, size_t nbyte);
 char		*ft_strdup(const char *s1);
 char		*ft_malloc(const char *s1);
